+++ include/unity/scopes/ScopeMetadata.h
@@ -20,6 +20,7 @@
 #define UNITY_SCOPES_SCOPEMETADATA_H
 
 #include <unity/scopes/Scope.h>
+#include <vector>
 
 namespace unity
 {
@@ -182,6 +183,18 @@ public:
     \return True if this scope wants location data.
     */
     bool location_data_needed() const;  // optional (default = false)
+
+    /**
+    \brief Return the list of scope identifiers aggregated by this scope.
+
+    The list returned by this method comes from the .ini file.
+    The scope author must ensure that it contains all scopes that an aggregator
+    might collect results from. This list may contain scopes that are not currently
+    installed and are optional for proper functioning of the aggregator scope.
+
+    \return The list of scopes ids aggregated by this scope.
+    */
+    std::vector<std::string> child_scope_ids() const;
 
 private:
     ScopeMetadata(std::unique_ptr<internal::ScopeMetadataImpl>);           // Instantiable only by ScopeMetadataImpl