+++ include/unity/scopes/internal/ScopeMetadataImpl.h
@@ -58,6 +58,7 @@ public:
     ScopeMetadata::ResultsTtlType results_ttl_type() const;      // optional (default: none)
     VariantArray settings_definitions() const;                   // optional (default: empty array)
     bool location_data_needed() const;                           // optional (default: false)
+    std::vector<std::string> child_scope_ids() const;            // optional (default: empty array)
 
     void set_scope_id(std::string const& scope_id);
     void set_proxy(ScopeProxy const& proxy);
@@ -74,6 +75,7 @@ public:
     void set_results_ttl_type(ScopeMetadata::ResultsTtlType results_ttl);
     void set_settings_definitions(VariantArray const& settings_definitions);
     void set_location_data_needed(bool location_data_needed);
+    void set_child_scope_ids(std::vector<std::string> const& ids);
 
     VariantMap serialize() const;
     void deserialize(VariantMap const& var);
@@ -98,6 +100,7 @@ private:
     ScopeMetadata::ResultsTtlType results_ttl_type_;
     std::unique_ptr<VariantArray> settings_definitions_;  // Optional, hence a pointer
     std::unique_ptr<bool> location_data_needed_;          // Optional, hence a pointer
+    std::vector<std::string> child_scope_ids_;
 };
 
 } // namespace internal