+++ scoperegistry/scoperegistry.cpp
@@ -315,6 +315,7 @@ void add_local_scope(RegistryObject::SPtr const& registry,
     mi->set_author(sc.author());
     mi->set_invisible(sc.invisible());
     mi->set_appearance_attributes(sc.appearance_attributes());
+    mi->set_child_scope_ids(sc.child_scope_ids());
 
     // Prepend scope_dir to pageheader logo path if logo path is relative.
     // TODO: Once we have type-safe parsing in the config files, remove