+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release notes
 
 Changes in version 0.6.8
 ========================
+  - Added support for ChildScopes scope .ini option to list scopes ids of aggregated scopes.
   - Replaced dbus-send with "list updated" pub/sub to invalidate smart scopes.
 
 Changes in version 0.6.7